+++ TidalPy/radial_solver/numerical/interfaces/liquid_liquid.py
@@ -1,5 +1,8 @@
 """ Functions to calculate the initial conditions for an overlying liquid layer above another liquid layer.
 
+For liquid-liquid layer interfaces: all the radial functions are continuous expect for if you are moving
+from a dynamic layer to static layer or vice versa.
+
 References
 ----------
 S74   : Saito (1974; J. Phy. Earth; DOI: 10.4294/jpe1952.22.123)
@@ -11,17 +14,14 @@ import numpy as np
 from TidalPy.constants import G
 from TidalPy.utilities.performance import njit, nbList
 
-from ..initial_conditions.initial_solution_dynamic import LiquidDynamicGuess, SolidDynamicGuess
-from ..initial_conditions.initial_solution_static import LiquidStaticGuess, SolidStaticGuess
+from ..initial_conditions.initial_solution_dynamic import LiquidDynamicGuess
+from ..initial_conditions.initial_solution_static import LiquidStaticGuess
 
-
-# For liquid-liquid layer interfaces: all the radial functions are continuous expect for if you are moving
-#    from a dynamic to static or vice versa.
 
 @njit(cacheable=True)
 def both_dynamic(liquid_layer_ys: LiquidDynamicGuess) -> LiquidDynamicGuess:
-    """ Calculated the starting values for the radial functions at the bottom of a liquid layer that is above another
-    liquid surface. Assumes dynamic tides in both solid layers.
+    """ Find the starting values for the radial functions at the bottom of a liquid layer that is above another
+    liquid layer. Assumes dynamic tides in both layers.
 
     References
     ----------
@@ -30,20 +30,21 @@ def both_dynamic(liquid_layer_ys: LiquidDynamicGuess) -> LiquidDynamicGuess:
     Parameters
     ----------
     liquid_layer_ys : LiquidDynamicGuess
-        The solution for the radial functions in the layer below, this function assumes a dynamic liquid lower layer
+        The solution for the radial functions in the layer below.
+        This function assumes a lower layer that is liquid and dynamic.
 
     Returns
     -------
-    base_liquid_ys : LiquidDynamicGuess
-        The base (initial) solutions used to calculate the radial functions in the solid layer.
-        For the assumptions used in this model there will be two independent solutions.
+    initial_solutions_liquid : LiquidDynamicGuess
+        The base (initial) solutions used to calculate the radial functions in the upper layer.
+        For this function's assumptions, there will be two independent solutions for the upper layer.
     """
 
-    base_liquid_ys = [
-        np.ascontiguousarray(liquid_layer_ys[0][:, -1]),
-        np.ascontiguousarray(liquid_layer_ys[1][:, -1])
-        ]
-    return nbList(base_liquid_ys)
+    initial_solutions_liquid = nbList([
+        np.ascontiguousarray(liquid_layer_ys[0]),
+        np.ascontiguousarray(liquid_layer_ys[1])
+        ])
+    return initial_solutions_liquid
 
 
 @njit(cacheable=True)
@@ -51,8 +52,8 @@ def static_dynamic(
         liquid_layer_ys: LiquidStaticGuess,
         interface_gravity: float, liquid_density: float, G_to_use: float = G
         ) -> LiquidDynamicGuess:
-    """ Calculated the starting values for the radial functions at the bottom of a liquid layer that is above another
-    liquid surface. Assumes dynamic tides in the upper layer and static tides in the lower.
+    """ Find the starting values for the radial functions at the bottom of a liquid layer that is above another
+    liquid layer. Assumes static tides in the lower layer and dynamic tides the upper.
 
     References
     ----------
@@ -61,20 +62,21 @@ def static_dynamic(
     Parameters
     ----------
     liquid_layer_ys : LiquidStaticGuess
-        The solution for the radial functions in the layer below, this function assumes a static liquid lower layer
+        The solution for the radial functions in the layer below.
+        This function assumes a lower layer that is liquid and static.
     interface_gravity : float
-        Acceleration due to gravity at the interface [m s-2]
+        Acceleration due to gravity at the interface [m s-2].
     liquid_density : float
-        The density at the top of the liquid layer (liquid's density) [kg m-3]
+        The density at the top of the liquid layer (liquid's density) [kg m-3].
         For this method we assume that the density provided is for the _static_ layer.
     G_to_use : float = G
         Gravitational constant. Provide a non-dimensional version if the rest of the inputs are non-dimensional.
 
     Returns
     -------
-    base_liquid_ys : LiquidDynamicGuess
-        The base (initial) solutions used to calculate the radial functions in the solid layer.
-        For the assumptions used in this model there will be two independent solutions.
+    initial_solutions_liquid : LiquidDynamicGuess
+        The base (initial) solutions used to calculate the radial functions in the upper layer.
+        For this function's assumptions, there will be two independent solutions for the upper layer.
     """
 
     # JPR decided to follow a similar approach as Eq. 20 in S74:
@@ -82,40 +84,41 @@ def static_dynamic(
     #   The upper dynamic liquid layer is treated like the solid layer in Eq. 20 except
     #    that y_3 is undefined as is "set 3" solution mentioned in that text.
 
-    # For a dynamic upper liquid layer there will be three independent solutions that we need an initial guess for.
-    base_liquid_list = nbList()
-
     # For a static lower liquid layer there will be one independent solutions at the top of the layer
     #   however for consistency this single solution is still stored in a tuple (of size 1) so we need to pull it out.
-    liquid_static_sol = liquid_layer_ys[0][:, -1]
+    liquid_static_sol = liquid_layer_ys[0]
+
+    # For an upper layer that is liquid and dynamic there will be two independent solutions that need an initial guess.
+    # For a dynamic liquid layer there will be four y values used.
+    initial_solutions_liquid = nbList([
+        np.empty(4, dtype=np.complex128),
+        np.empty(4, dtype=np.complex128)
+        ])
 
     for solution in (0, 1):
-        # For a dynamic liquid layer there will be four y values used.
-        solution_values = np.empty(4, dtype=liquid_static_sol.dtype)
+        solution_ys = initial_solutions_liquid[solution]
 
         if solution == 0:
             # y_1_dynamic = 0
-            solution_values[0] = 0.
+            solution_ys[0] = 0.
             # y_2_dynamic = -rho * y_5_static
-            solution_values[1] = -liquid_density * liquid_static_sol[0]
+            solution_ys[1] = -liquid_density * liquid_static_sol[0]
             # y_5_dynamic = y_5_static
-            solution_values[2] = liquid_static_sol[0]
+            solution_ys[2] = liquid_static_sol[0]
             # y_6_dynamic = y_7_static + (4 pi G rho / g) y_5_static
-            solution_values[3] = liquid_static_sol[1] + \
-                                 (4. * np.pi * G_to_use * liquid_density / interface_gravity) * liquid_static_sol[0]
+            solution_ys[3] = liquid_static_sol[1] + \
+                             (4. * np.pi * G_to_use * liquid_density / interface_gravity) * liquid_static_sol[0]
         else:
             # y_1_dynamic = 1.
-            solution_values[0] = 1.
+            solution_ys[0] = 1.
             # y_2_dynamic = rho * g * y_1_dynamic
-            solution_values[1] = liquid_density * interface_gravity * solution_values[0]
+            solution_ys[1] = liquid_density * interface_gravity * solution_ys[0]
             # y_5_dynamic = 0.
-            solution_values[2] = 0.
+            solution_ys[2] = 0.
             # y_6_dynamic = -4 pi G rho y_1_dynamic
-            solution_values[3] = -4. * np.pi * G_to_use * liquid_density * solution_values[0]
+            solution_ys[3] = -4. * np.pi * G_to_use * liquid_density * solution_ys[0]
 
-        base_liquid_list.append(solution_values)
-
-    return base_liquid_list
+    return initial_solutions_liquid
 
 
 @njit(cacheable=True)
@@ -123,8 +126,8 @@ def dynamic_static(
         liquid_layer_ys: LiquidDynamicGuess,
         interface_gravity: float, liquid_density: float, G_to_use: float = G
         ) -> LiquidStaticGuess:
-    """ Calculated the starting values for the radial functions at the bottom of a liquid layer that is above another
-    liquid surface. Assumes dynamic tides in the upper layer and static tides in the lower.
+    """ Find the starting values for the radial functions at the bottom of a liquid layer that is above another
+    liquid layer. Assumes static tides in the upper layer and dynamic tides the lower.
 
     References
     ----------
@@ -134,20 +137,21 @@ def dynamic_static(
     Parameters
     ----------
     liquid_layer_ys : LiquidDynamicGuess
-        The solution for the radial functions in the layer below, this function assumes a dynamic liquid lower layer
+        The solution for the radial functions in the layer below.
+        This function assumes a lower layer that is liquid and dynamic.
     interface_gravity : float
-        Acceleration due to gravity at the interface [m s-2]
+        Acceleration due to gravity at the interface [m s-2].
     liquid_density : float
-        The density at the top of the liquid layer (liquid's density) [kg m-3]
+        The density at the top of the liquid layer (liquid's density) [kg m-3].
         For this method we assume that the density provided is for the _static_ layer.
     G_to_use : float = G
         Gravitational constant. Provide a non-dimensional version if the rest of the inputs are non-dimensional.
 
     Returns
     -------
-    base_liquid_ys : LiquidStaticGuess
-        The base (initial) solutions used to calculate the radial functions in the solid layer.
-        For the assumptions used in this model there will be one independent solutions.
+    initial_solutions_liquid : LiquidStaticGuess
+        The base (initial) solutions used to calculate the radial functions in the upper layer.
+        For this function's assumptions, there will be one independent solutions for the upper layer.
     """
 
     # JPR decided to follow a similar approach as Eq. 20 in S74:
@@ -156,13 +160,13 @@ def dynamic_static(
     #    that y_3 is undefined as is "set 3" solution mentioned in that text.
 
     # For a static liquid layer there will be one independent solution with 2 y's
-    base_liquid_ys = np.empty(2, dtype=liquid_layer_ys[0].dtype)
+    initial_solutions_liquid = np.empty(2, dtype=np.complex128)
 
     # lambda_j = (y_2j - rho * ( g * y_1j - y_5j))
-    lambda_1 = liquid_layer_ys[0][1, -1] - \
-               liquid_density * (interface_gravity * liquid_layer_ys[0][0, -1] - liquid_layer_ys[0][4, -1])
-    lambda_2 = liquid_layer_ys[1][1, -1] - \
-               liquid_density * (interface_gravity * liquid_layer_ys[1][0, -1] - liquid_layer_ys[1][4, -1])
+    lambda_1 = liquid_layer_ys[0][1] - \
+               liquid_density * (interface_gravity * liquid_layer_ys[0][0] - liquid_layer_ys[0][4])
+    lambda_2 = liquid_layer_ys[1][1] - \
+               liquid_density * (interface_gravity * liquid_layer_ys[1][0] - liquid_layer_ys[1][4])
 
     # Set the first coefficient to 1. It will be solved for later on during the collapse phase.
     coeff_1 = 1.
@@ -170,21 +174,21 @@ def dynamic_static(
     coeff_2 = -(lambda_1 / lambda_2) * coeff_1
 
     y_7_const = (4. * np.pi * G_to_use / interface_gravity)
-    y_7_IC_0 = liquid_layer_ys[0][5, -1] + y_7_const * liquid_layer_ys[0][1, -1]
-    y_7_IC_1 = liquid_layer_ys[1][5, -1] + y_7_const * liquid_layer_ys[1][1, -1]
+    y_7_IC_0 = liquid_layer_ys[0][5] + y_7_const * liquid_layer_ys[0][1]
+    y_7_IC_1 = liquid_layer_ys[1][5] + y_7_const * liquid_layer_ys[1][1]
 
     # y^liq_5 = C^sol_1 * y^sol_5,1 + C^sol_2 * y^sol_5,2
-    base_liquid_ys[0] = coeff_1 * liquid_layer_ys[0][4, -1] + coeff_2 * liquid_layer_ys[1][4, -1]
+    initial_solutions_liquid[0] = coeff_1 * liquid_layer_ys[0][4] + coeff_2 * liquid_layer_ys[1][4]
     # y^liq_7 = C^sol_1 * y^sol_7,1 + C^sol_2 * y^sol_7,2
-    base_liquid_ys[1] = coeff_1 * y_7_IC_0 + coeff_2 * y_7_IC_1
+    initial_solutions_liquid[1] = coeff_1 * y_7_IC_0 + coeff_2 * y_7_IC_1
 
-    return nbList([base_liquid_ys])
+    return nbList([initial_solutions_liquid])
 
 
 @njit(cacheable=True)
 def both_static(liquid_layer_ys: LiquidStaticGuess) -> LiquidStaticGuess:
-    """ Calculated the starting values for the radial functions at the bottom of a liquid layer that is above another
-    liquid surface. Assumes static tides in both solid layers.
+    """ Find the starting values for the radial functions at the bottom of a liquid layer that is above another
+    liquid layer. Assumes static tides in both the upper and lower layers.
 
     References
     ----------
@@ -193,16 +197,16 @@ def both_static(liquid_layer_ys: LiquidStaticGuess) -> LiquidStaticGuess:
     Parameters
     ----------
     liquid_layer_ys : LiquidStaticGuess
-        The solution for the radial functions in the layer below, this function assumes a static liquid lower layer
+        The solution for the radial functions in the layer below.
+        This function assumes a lower layer that is liquid and static.
 
     Returns
     -------
-    base_liquid_ys : LiquidStaticGuess
-        The base (initial) solutions used to calculate the radial functions in the solid layer.
-        For the assumptions used in this model there will be one independent solution.
+    initial_solutions_liquid : LiquidStaticGuess
+        The base (initial) solutions used to calculate the radial functions in the upper layer.
+        For this function's assumptions, there will be one independent solutions for the upper layer.
     """
 
-    base_liquid_ys = liquid_layer_ys[0][:, -1]
-    base_liquid_ys = nbList([np.ascontiguousarray(base_liquid_ys)])
+    initial_solutions_liquid = nbList([np.ascontiguousarray(liquid_layer_ys[0])])
 
-    return base_liquid_ys
+    return initial_solutions_liquid